requirements.txt
allennlp==2.0.1
allennlp-models==2.0.1
attrs==20.3.0
bitarray==1.6.3
bitstring==3.1.9
blis==0.7.5
boto3==1.20.29
botocore==1.23.29
cached-property==1.5.2
camelsplit==1.0.1
catalogue==1.0.0
charset-normalizer==2.0.10
click==7.1.2
conllu==4.3
cycler==0.11.0
cymem==2.0.6
datasketch==1.5.3
elasticsearch==7.10.1
filelock==3.0.12
ftfy==6.0.3
future==0.18.2
h5py==3.1.0
idna==3.3
importlib-metadata==3.4.0
iniconfig==1.1.1
isodate==0.6.0
jmespath==0.10.0
joblib==1.0.0
jsonnet==0.18.0
jsonpickle==2.0.0
kiwisolver==1.3.1
lmdb==1.3.0
logger==1.4
matplotlib==3.3.4
more-itertools==8.12.0
murmurhash==1.0.6
NearPy==1.0.0
nltk==3.5
overrides==3.1.0
packaging==20.9
Pillow==8.4.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.6
protobuf==3.19.1
py==1.10.0
py-rouge==1.1
py4j==0.10.9
pyparsing==2.4.7
pyspark==3.0.1
pyspark-stubs==3.0.0.post1
pytest==6.2.2
PyYAML==5.4.1
rdflib==5.0.0
regex==2020.11.13
requests==2.27.1
s3transfer==0.5.0
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.5.4
sentencepiece==0.1.96
spacy==2.3.7
SPARQLWrapper==1.8.5
srsly==1.0.5
tensorboardX==2.4.1
thinc==7.4.5
threadpoolctl==2.1.0
tokenizers==0.9.4
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
transformers==4.2.2
typing-extensions==3.7.4.3
wasabi==0.9.0
wcwidth==0.2.5
word2number==1.1
zipp==3.4.0
astor==0.8.1
certifi==2021.10.8
graphviz==0.17
neo4j==4.3.6
numpy==1.21.4
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
staticfg==0.9.5
urllib3==1.26.7
